Update and rename clipoptionslist.php to clip_options_list.php
